+++ packages/douyu-cli/src/core/index.ts
@@ -5,16 +5,9 @@ import axios from "axios";
 import M3U8Downloader from "@renmu/m3u8-downloader";
 
 import { sanitizeFileName } from "../utils/index.js";
-import {
-  getVideoDanmu,
-  getStreamUrls,
-  getVideos,
-  parseVideo,
-  getReplayList,
-  convert2Xml,
-} from "douyu-api";
+import { video as videoAPI, convert2Xml } from "douyu-api";
 import up from "./up.js";
-import { readData, pushData, deleteData, readConfig } from "./config.js";
+import { readData, pushData, deleteData } from "./config.js";
 import logger from "../utils/log.js";
 
 import type { Video, streamType } from "douyu-api";
@@ -28,6 +21,10 @@ interface DownloadOptions {
   video?: boolean;
   ffmpegBinPath?: string;
   concurrency?: number;
+  startIndex?: number;
+  endIndex?: number;
+  startTime?: number;
+  endTime?: number;
 }
 
 /**
@@ -39,14 +36,14 @@ export const subscribe = async (options: DownloadOptions) => {
 
   let videoIds: string[] = [];
   for (const up of upList) {
-    const replayList = await getReplayList({
+    const replayList = await videoAPI.getReplayList({
       up_id: up.upId,
       page: 1,
       limit: 1,
     });
     for (const replay of replayList.list) {
       for (const video of replay.video_list) {
-        const videos = await getVideos(video.hash_id, up.upId);
+        const videos = await videoAPI.getVideos(video.hash_id, up.upId);
         videoIds.push(...videos.list.map(item => item.hash_id));
       }
     }
@@ -88,13 +85,13 @@ export const downloadVideos = async (
     danmaku: false,
   }
 ) => {
-  const videoData = await parseVideo(buildVideoUrl(videoId));
+  const videoData = await videoAPI.parseVideo(buildVideoUrl(videoId));
   const downloadDir = opts.dir!;
 
   if (opts.all) {
-    const res = await getVideos(videoId, videoData.ROOM.up_id);
+    const res = await videoAPI.getVideos(videoId, videoData.ROOM.up_id);
     for (const video of res.list) {
-      const videoData = await parseVideo(buildVideoUrl(video.hash_id));
+      const videoData = await videoAPI.parseVideo(buildVideoUrl(video.hash_id));
       const name = sanitizeFileName(video.title);
       const output = path.join(downloadDir, `${name}.mp4`);
       // console.log(JSON.stringify(videoData, null, 2));
@@ -205,7 +202,7 @@ export const downloadVideo = async (
  * 获取最高清晰度的视频流
  */
 const getStream = async (data: string, streamType?: streamType) => {
-  const res = await getStreamUrls(data);
+  const res = await videoAPI.getStreamUrls(data);
   const streams = Object.values(res.thumb_video);
   if (streams.length === 0) {
     throw new Error("没有找到视频流");
@@ -253,6 +250,10 @@ const saveVideo = async (
         concurrency: opts.concurrency || 10,
         ffmpegPath: opts.ffmpegBinPath,
         segmentsDir,
+        startIndex: opts.startIndex,
+        endIndex: opts.endIndex,
+        startTime: opts.startTime,
+        endTime: opts.endTime,
       },
       data => {
         const percentage = Math.floor((data.downloaded / data.total) * 100);
@@ -274,7 +275,7 @@ export async function saveDanmu(vid: string, output: string, video: Video) {
     logger.info(`弹幕文件已存在，跳过下载`);
     return;
   }
-  const items = await getVideoDanmu(vid);
+  const items = await videoAPI.getVideoDanmu(vid);
   let room_title = video.ROOM.name;
   if (room_title.startsWith("【") && room_title.split("：").length > 1) {
     room_title = room_title.split("：").slice(1).join("：");
@@ -304,6 +305,10 @@ export const downloadHLS = (
     concurrency?: number;
     ffmpegPath?: string;
     segmentsDir: string;
+    startIndex?: number;
+    endIndex?: number;
+    startTime?: number;
+    endTime?: number;
   },
   onProgress?: (data: { downloaded: number; total: number }) => void
 ) => {